# bbox-tile-server/src/openapi.yaml
openapi: 3.0.0
info:
  title: 'Building Blocks specified in OGC API - Tiles - Part 1: Core'
  description: 'Draft example of OGC API - Tiles - Part 1 OpenAPI docs. This document is not an OGC Standard. This document is distributed for review and comment. This document is subject to change without notice and may not be referred to as an OGC Standard. Recipients of this document are invited to submit, with their comments, notification of any relevant patent rights of which they are aware and to provide supporting documentation.'
  version: 0.0.2
  contact:
    name: "Open Geospatial Consortium (OGC)"
    email: standards-team@ogc.org
    url: 'https://www.ogc.org/contacts'
  license:
    name: OGC License
    url: http://www.opengeospatial.org/legal/
  x-logo:
    url: 'https://www.ogc.org/pub/www/files/OGC_Logo_2D_Blue_x_0_0.png'
servers:
- description: Example OGC API - Tiles server
  url: https://maps.ecere.com/ogcapi
paths:
...
  "/map/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}":
    get:
      tags:
      - Map Tiles
      summary: Retrieve a default map tile of the whole dataset
      operationId: ".dataset.map.getTile"
      parameters:
      - "$ref": "#/components/parameters/tileMatrix"
      - "$ref": "#/components/parameters/tileRow"
      - "$ref": "#/components/parameters/tileCol"
      # - "$ref": "#/components/parameters/datetime"
      # - "$ref": "#/components/parameters/collections"
      # - "$ref": "#/components/parameters/subset"
      # - "$ref": "#/components/parameters/crs"
      # - "$ref": "#/components/parameters/subset-crs"
      # - "$ref": "#/components/parameters/bgcolor"
      # - "$ref": "#/components/parameters/transparent"
      - "$ref": "#/components/parameters/tileMatrixSetId"
      # - "$ref": "#/components/parameters/f-mapTile"
      responses:
        '200':
          "$ref": "#/components/responses/MapTile"
        '204':
          "$ref": "#/components/responses/EmptyTile"
        '404':
          "$ref": "#/components/responses/NotFound"
        '406':
          "$ref": "#/components/responses/NotAcceptable"
        '500':
          "$ref": "#/components/responses/ServerError"
  # "/styles/{styleId}/map/tiles":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve the list of styled map tilesets for the whole dataset
  #     operationId: ".dataset.style.map.getTileSetsList"
  #     parameters:
  #     - "$ref": "#/components/parameters/styleId"
  #     - "$ref": "#/components/parameters/f-metadata"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/TileSetsList"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/styles/{styleId}/map/tiles/{tileMatrixSetId}":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve a styled map tileset of the whole dataset for the specified
  #       tiling scheme (tile matrix set)
  #     operationId: ".dataset.style.map.getTileSet"
  #     parameters:
  #     - "$ref": "#/components/parameters/styleId"
  #     - "$ref": "#/components/parameters/collections"
  #     - "$ref": "#/components/parameters/tileMatrixSetId"
  #     - "$ref": "#/components/parameters/f-metadata"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/TileSet"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/styles/{styleId}/map/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve a styled map tiles
  #     operationId: ".dataset.style.map.getTile"
  #     parameters:
  #     - "$ref": "#/components/parameters/tileMatrix"
  #     - "$ref": "#/components/parameters/tileRow"
  #     - "$ref": "#/components/parameters/tileCol"
  #     - "$ref": "#/components/parameters/datetime"
  #     - "$ref": "#/components/parameters/collections"
  #     - "$ref": "#/components/parameters/subset"
  #     - "$ref": "#/components/parameters/crs"
  #     - "$ref": "#/components/parameters/subset-crs"
  #     - "$ref": "#/components/parameters/bgcolor"
  #     - "$ref": "#/components/parameters/transparent"
  #     - "$ref": "#/components/parameters/styleId"
  #     - "$ref": "#/components/parameters/tileMatrixSetId"
  #     - "$ref": "#/components/parameters/f-mapTile"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/MapTile"
  #       '204':
  #         "$ref": "#/components/responses/EmptyTile"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/collections/{collectionId}/map/tiles":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve a list of all map tilesets for specified collection.
  #     operationId: ".collection.map.getTileSetsList"
  #     parameters:
  #     - "$ref": "#/components/parameters/collectionId-all"
  #     - "$ref": "#/components/parameters/f-metadata"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/TileSetsList"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/collections/{collectionId}/map/tiles/{tileMatrixSetId}":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve a map tile set metadata for the specified collection and tiling
  #       scheme (tile matrix set)
  #     operationId: ".collection.map.getTileSet"
  #     parameters:
  #     - "$ref": "#/components/parameters/collectionId-all"
  #     - "$ref": "#/components/parameters/collections"
  #     - "$ref": "#/components/parameters/tileMatrixSetId"
  #     - "$ref": "#/components/parameters/f-metadata"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/TileSet"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/collections/{collectionId}/map/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve a map tile from the specified collection
  #     operationId: ".collection.map.getTile"
  #     parameters:
  #     - "$ref": "#/components/parameters/tileMatrix"
  #     - "$ref": "#/components/parameters/tileRow"
  #     - "$ref": "#/components/parameters/tileCol"
  #     - "$ref": "#/components/parameters/datetime"
  #     - "$ref": "#/components/parameters/collectionId-all"
  #     - "$ref": "#/components/parameters/collections"
  #     - "$ref": "#/components/parameters/subset"
  #     - "$ref": "#/components/parameters/crs"
  #     - "$ref": "#/components/parameters/subset-crs"
  #     - "$ref": "#/components/parameters/bgcolor"
  #     - "$ref": "#/components/parameters/transparent"
  #     - "$ref": "#/components/parameters/tileMatrixSetId"
  #     - "$ref": "#/components/parameters/f-mapTile"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/MapTile"
  #       '204':
  #         "$ref": "#/components/responses/EmptyTile"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/collections/{collectionId}/styles/{styleId}/map/tiles":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve a list of styled map tilesets for the specified collection
  #     operationId: ".collection.style.map.getTileSetsList"
  #     parameters:
  #     - "$ref": "#/components/parameters/collectionId-all"
  #     - "$ref": "#/components/parameters/f-metadata"
  #     - "$ref": "#/components/parameters/styleId-collection"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/TileSetsList"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/collections/{collectionId}/styles/{styleId}/map/tiles/{tileMatrixSetId}":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve the map tileset metadata for the specified collection, style
  #       and tiling scheme (tile matrix set).
  #     operationId: ".collection.style.map.getTileSet"
  #     parameters:
  #     - "$ref": "#/components/parameters/collectionId-all"
  #     - "$ref": "#/components/parameters/collections"
  #     - "$ref": "#/components/parameters/styleId-collection"
  #     - "$ref": "#/components/parameters/tileMatrixSetId"
  #     - "$ref": "#/components/parameters/f-metadata"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/TileSet"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
  # "/collections/{collectionId}/styles/{styleId}/map/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}":
  #   get:
  #     tags:
  #     - Map Tiles
  #     summary: Retrieve a map tile for a specified collection and style
  #     operationId: ".collection.style.map.getTile"
  #     parameters:
  #     - "$ref": "#/components/parameters/tileMatrix"
  #     - "$ref": "#/components/parameters/tileRow"
  #     - "$ref": "#/components/parameters/tileCol"
  #     - "$ref": "#/components/parameters/datetime"
  #     - "$ref": "#/components/parameters/collectionId-all"
  #     - "$ref": "#/components/parameters/collections"
  #     - "$ref": "#/components/parameters/subset"
  #     - "$ref": "#/components/parameters/crs"
  #     - "$ref": "#/components/parameters/subset-crs"
  #     - "$ref": "#/components/parameters/bgcolor"
  #     - "$ref": "#/components/parameters/transparent"
  #     - "$ref": "#/components/parameters/styleId-collection"
  #     - "$ref": "#/components/parameters/tileMatrixSetId"
  #     - "$ref": "#/components/parameters/f-mapTile"
  #     responses:
  #       '200':
  #         "$ref": "#/components/responses/MapTile"
  #       '204':
  #         "$ref": "#/components/responses/EmptyTile"
  #       '404':
  #         "$ref": "#/components/responses/NotFound"
  #       '406':
  #         "$ref": "#/components/responses/NotAcceptable"
  #       '500':
  #         "$ref": "#/components/responses/ServerError"
components:
  schemas:
...
    link:
      type: object
      required:
      - href
      - rel
      properties:
        href:
          type: string
          description: Supplies the URI to a remote resource (or resource fragment).
          example: http://data.example.com/buildings/123
        rel:
          type: string
          description: The type or semantics of the relation.
          example: alternate
        type:
          type: string
          description: A hint indicating what the media type of the result of dereferencing
            the link should be.
          example: application/geo+json
        templated:
          type: boolean
          description: This flag set to true if the link is a URL template.
        hreflang:
          type: string
          description: A hint indicating what the language of the result of dereferencing
            the link should be.
          example: en
        title:
          type: string
          description: Used to label the destination of a link such that it can be
            used as a human-readable identifier.
          example: Trierer Strasse 70, 53115 Bonn
        length:
          type: integer
...
    exception:
      title: Exception Schema
      description: JSON schema for exceptions based on RFC 7807
      type: object
      required:
      - type
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    crs:
      title: CRS
      oneOf:
      - description: Simplification of the object into a url if the other properties
          are not present
        type: string
      - type: object
        oneOf:
        - required:
          - uri
          properties:
            uri:
              description: Reference to one coordinate reference system (CRS)
              type: string
              format: uri
        - required:
          - wkt
          properties:
            wkt:
              description: A string defining the CRS using the JSON encodng for Well
                Known Text
              type: object
        - required:
          - referenceSystem
          properties:
            referenceSystem:
              description: A reference system data structure as defined in the MD_ReferenceSystem
                of the ISO 19115
              type: object
    dataType:
      oneOf:
      - type: string
      - type: string
        enum:
        - map
        - vector
        - coverage
...
    tileSet-item:
      title: Tile Set Metadata item
      description: A minimal tileset element for use within a list of tilesets linking
        to full description of those tilesets.
      type: object
      required:
      - dataType
      - links
      - crs
      properties:
        title:
          description: A title for this tileset
          type: string
        dataType:
          allOf:
          - description: Type of data represented in the tileset
          - "$ref": "#/components/schemas/dataType"
        crs:
          allOf:
          - description: Coordinate Reference System (CRS)
          - "$ref": "#/components/schemas/crs"
        tileMatrixSetURI:
          description: Reference to a Tile Matrix Set on an offical source for Tile
            Matrix Sets such as the OGC NA definition server (http://www.opengis.net/def/tms/).
            Required if the tile matrix set is registered on an open official source.
          type: string
          format: uri
        links:
          description: Links to related resources. A 'self' link to the tileset as
            well as a 'http://www.opengis.net/def/rel/ogc/1.0/tiling-scheme' link
            to a definition of the TileMatrixSet are required.
          type: array
          items:
            "$ref": "#/components/schemas/link"
...
    all-collections:
      type: string
      enum:
      - blueMarble
      - NaturalEarth:raster:HYP_HR_SR_OB_DR
      - NaturalEarth:cultural:ne_10m_admin_0_countries
      - NaturalEarth:physical:bathymetry
      - SRTM_ViewFinderPanorama
      - HRDEM-Ottawa
      - HRDEM-RedRiver
...
    styles:
      type: string
      enum:
      - default
    tileMatrixSets:
      type: string
      enum:
      - WebMercatorQuad
      - WorldCRS84Quad
      - GNOSISGlobalGrid
      - WorldMercatorWGS84Quad
  parameters:
    f-metadata:
      name: f
      in: query
      description: The format of the response. If no value is provided, the accept
        header is used to determine the format. Accepted values are 'json' or 'html'.
      required: false
      schema:
        type: string
        enum:
        - json
        - html
      style: form
      explode: false
    collectionId-all:
      name: collectionId
      in: path
      description: Local identifier of a collection
      required: true
      schema:
        "$ref": "#/components/schemas/all-collections"
...
    collections:
      name: collections
      in: query
      style: form
      description: The collections that should be included in the response. The parameter
        value is a comma-separated list of collection identifiers. If the parameters
        is missing, some or all collections will be included. The collection will
        be rendered in the order specified, with the last one showing on top, unless
        the priority is overridden by styling rules.
      required: false
      explode: false
      schema:
        type: array
        items:
          "$ref": "#/components/schemas/all-collections"
...
    datetime:
      name: datetime
      in: query
      description: |-
        Either a date-time or an interval, half-bounded or bounded. Date and time expressions
        adhere to RFC 3339. Half-bounded intervals are expressed using double-dots.

        Examples:

        * A date-time: "2018-02-12T23:20:50Z"
        * A bounded interval: "2018-02-12T00:00:00Z/2018-03-18T12:31:12Z"
        * Half-bounded intervals: "2018-02-12T00:00:00Z/.." or "../2018-03-18T12:31:12Z"

        Only features that have a temporal property that intersects the value of
        `datetime` are selected.

        If a feature has multiple temporal properties, it is the decision of the
        server whether only a single temporal property is used to determine
        the extent or all relevant temporal properties.
      required: false
      schema:
        type: string
      style: form
      explode: false
    subset:
      name: subset
      in: query
      description: |
        Retrieve only part of the data by slicing or trimming along one or more axis
        For trimming: {axisAbbrev}({low}:{high}) (preserves dimensionality)
           An asterisk (`*`) can be used instead of {low} or {high} to indicate the minimum/maximum value.
        For slicing:  {axisAbbrev}({value})      (reduces dimensionality)
      style: form
      explode: false
      required: false
      schema:
        type: array
        items:
          type: string
    crs:
      name: crs
      in: query
      description: reproject the output to the given crs
      required: false
      style: form
      explode: true
      schema:
        type: string
    subset-crs:
      name: subset-crs
      in: query
      description: crs for the specified subset
      required: false
      style: form
      explode: true
      schema:
        type: string
    tileMatrix:
      name: tileMatrix
      in: path
      description: |-
        Identifier selecting one of the scales defined in the TileMatrixSet and representing the scaleDenominator the tile. For example,
        Ireland is fully within the Tile at WebMercatorQuad tileMatrix=5, tileRow=10 and tileCol=15.
      required: true
      schema:
        type: string
      example: '5'
    tileRow:
      name: tileRow
      in: path
      description: Row index of the tile on the selected TileMatrix. It cannot exceed
        the MatrixWidth-1 for the selected TileMatrix. For example, Ireland is fully
        within the Tile at WebMercatorQuad tileMatrix=5, tileRow=10 and tileCol=15.
      required: true
      schema:
        minimum: 0
        type: integer
      example: 10
    tileCol:
      name: tileCol
      in: path
      description: Column index of the tile on the selected TileMatrix. It cannot
        exceed the MatrixHeight-1 for the selected TileMatrix. For example, Ireland
        is fully within the Tile at WebMercatorQuad tileMatrix=5, tileRow=10 and tileCol=15.
      required: true
      schema:
        minimum: 0
        type: integer
      example: 15
    tileMatrixSetId:
      name: tileMatrixSetId
      in: path
      description: Identifier for a supported TileMatrixSet
      required: true
      allowEmptyValue: false
      schema:
        "$ref": "#/components/schemas/tileMatrixSets"
    f-mapTile:
      name: f
      in: query
      description: The format of the map tile response (e.g. png). Accepted values
        are 'png', 'jpg' or 'tiff' (GeoTIFF).
      required: false
      schema:
        type: string
        enum:
        - png
        - jpg
        - tiff
      style: form
      explode: false
...
    bgcolor:
      name: bgcolor
      in: query
      description: Web color name or hexadecimal 0x[AA]RRGGBB color value for the
        background color (default to 0x9C9C9C gray). If alpha is not specified, full
        opacity is assumed.
      required: false
      style: form
      explode: false
      schema:
        type: string
        default: '0xFFFFFF'
    transparent:
      name: transparent
      in: query
      description: Background transparency of map (default=true).
      required: false
      style: form
      explode: false
      schema:
        type: boolean
        default: true
    styleId:
      name: styleId
      in: path
      description: An identifier representing a specific style.
      required: true
      allowEmptyValue: false
      schema:
        "$ref": "#/components/schemas/styles"
    styleId-collection:
      name: styleId
      in: path
      description: An identifier representing a specific style.
      required: true
      allowEmptyValue: false
      schema:
        type: string
  responses:
    NotFound:
      description: The requested resource does not exist on the server. For example,
        a path parameter had an incorrect value.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/exception"
        text/html:
          schema:
            type: string
    NotAcceptable:
      description: Content negotiation failed. For example, the `Accept` header submitted
        in the request did not support any of the media types supported by the server
        for the requested resource.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/exception"
        text/html:
          schema:
            type: string
    ServerError:
      description: A server error occurred.
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/exception"
        text/html:
          schema:
            type: string
...
    TileSetsList:
      description: List of available tilesets.
      content:
        application/json:
          schema:
            type: object
            required:
            - tilesets
            properties:
              links:
                type: array
                items:
                  "$ref": "#/components/schemas/link"
              tilesets:
                type: array
                items:
                  "$ref": "#/components/schemas/tileSet-item"
        text/html:
          schema:
            type: string
    TileSet:
      description: Description of the tileset
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/tileSet"
        text/html:
          schema:
            type: string
    MapTile:
      description: A map tile image returned as a response.
      content:
        image/png:
          schema:
            type: string
            format: binary
        image/jpeg:
          schema:
            type: string
            format: binary
        image/tiff; application=geotiff:
          schema:
            type: string
            format: binary
...
    EmptyTile:
      description: No data available for this tile.
...